Slap InstalmentTest into separate Test files
