+++ data/account-test.txt
@@ -1,80 +1,2 @@
 INIT @ true
 BS @ 0.0
-INC @ 2000.0 @ TA Pay @ 9/10/2015
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 200.0 @ Deposit to OCBC @ 3/3/2018
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 200.0 @ Deposit to OCBC @ 3/3/2018
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 200.0 @ Deposit to OCBC @ 3/3/2018
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 200.0 @ Deposit to OCBC @ 3/3/2018
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 200.0 @ Deposit to OCBC @ 3/3/2018
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 200.0 @ Deposit to OCBC @ 3/3/2018
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 200.0 @ Deposit to OCBC @ 3/3/2018
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-INC @ 30.0 @ Initialize account: OCBC @ 27/7/2017
-EXP @ 1000.0 @ watch @ GS @ 26/10/2019
-EXP @ 200.0 @ Withdraw from OCBC @ withdraw from bank @ 3/3/2019
-EXP @ 30.0 @ Withdraw from OCBC @ bank @ 26/10/2019
-INS @ 100000.0 @ mortgage @ INS @ 12/12/2010 @ 200 @ 0.06
-BAN @ 30.0 @ OCBC @ 2017-07-27 @ 0.005
